+++ layouts/default.vue
@@ -24,7 +24,14 @@ html {
   box-sizing: border-box;
   margin: 0;
 }
-
+:root {
+  --bg-body: #1a1a2e;
+  --bg-btn-primary: #44cec0;
+  --bg-btn-danger: #e91e63;
+  --bg-btn-secondary: #3aa3ff;
+  --bg-btn-default: #3aa3ff;
+  --color-btn: #fff;
+}
 body {
   margin: 0;
   font-size: 16px;
